Create test for presence of links to invoices and items on merchant dashboard
